+++ package.json
@@ -47,12 +47,14 @@
   },
   "dependencies": {
     "bignumber.js": "^8",
+    "common-tags": "1",
     "deep-freeze": "^0.0.1",
     "es6-error": "^4",
     "lodash.isequal": "^4",
     "validator": "^10"
   },
   "devDependencies": {
+    "@types/common-tags": "^1.4.0",
     "@types/deep-freeze": "^0.1.1",
     "@types/jest": "^23",
     "@types/lodash.isequal": "^4",
