+++ src/Currency/Currency.test.ts
@@ -1,112 +1,162 @@
-import Currency from './Currency';
-import Money from '../Money/Money';
-import WrongInputError from '../errors/WrongInputError';
+import { USD, GBP, EUR } from '../constants/ISO_CURRENCIES';
+import Currency from '../Currency';
 import InvalidCurrencyError from '../errors/InvalidCurrencyError';
+import WrongInputError from '../errors/WrongInputError';
 
-describe('The "Currency" class: ', () => {
-  test('(new Currency("USD")).toString() === "USD"', () => {
-    expect((new Currency('USD')).toString()).toBe('USD');
+jest.mock('../CurrencyStore/data', () => ({
+  USD,
+  GBP,
+}));
+
+describe('Currency', () => {
+
+  test('can be initialized with the currency code of an already-registered currency', () => {
+    const currency = new Currency('USD');
+    expect(currency.settings).toEqual(USD);
   });
 
-  test('(new Currency("GBP")).decimalDigits === 2', () => {
-    expect((new Currency('GBP')).decimalDigits).toBe(2);
+  test('can be initialized with another Currency instance', () => {
+    const currency = new Currency('USD');
+    const currency2 = new Currency(currency);
+    expect(currency2.settings).toEqual(USD);
   });
 
-  test('(new Currency("EUR")).symbol === "€"', () => {
-    expect((new Currency('EUR')).symbol).toBe('€');
+  test('can be initialized directly with the currency settings', () => {
+    const currency = new Currency(EUR);
+    expect(currency.settings).toEqual(EUR);
   });
 
-  test('(new Currency("EUR")).pattern === "%ns%s%v"', () => {
-    expect((new Currency('CAD')).pattern).toBe('%ns%s%v');
-  });
-
-  test('(new Currency("GBP")).is(new Currency("GBP")) === true', () => {
-    expect((new Currency('GBP')).is(new Currency('GBP'))).toBe(true);
-  });
-
-  test('(new Currency("GBP")).is(new Currency("USD")) === false', () => {
-    expect((new Currency('GBP')).is(new Currency('USD'))).toBe(false);
-  });
-
-  test('clone should create a duplicate copy of a currency', () => {
-    const gbp = new Currency('GBP');
-    const gbpClone = gbp.clone();
-    expect(gbp).toEqual(gbpClone);
-  });
-
-  test('"formatter" option should do custom formatting', () => {
-    // wrapping the mock as the customFormmater gets frozen once passed in
-    const customFormatterInnerMock = jest.fn(() => 'custom format');
-    const customFormatter = (...args: any[]) => customFormatterInnerMock(...args);
-    const customGBP = new Currency({
-      ...Currency.getSettings('GBP'),
-      formatter: customFormatter,
+  describe('when partial settings are used', () => {
+    describe('and the currency is not registered', () => {
+      test('can be initialized with default built-in settings included', () => {
+        const currency = new Currency({ code: 'EUR' });
+        expect(currency.settings).toEqual({
+          code: 'EUR',
+          thousandsSeparator: ',',
+          decimalSeparator: '.',
+          decimalDigits: 2,
+          pattern: '%ns%s%v',
+          symbol: 'EUR',
+          formatter: undefined,
+          parser: undefined,
+        });
+      });
     });
-    const customGBPMoney = new Money('10.00', customGBP);
-    const formattedValue = customGBPMoney.format();
-    const {
-      formatter,
-      parser,
-      ...currencySettings
-    } = customGBP.settings;
 
-    expect(formattedValue).toBe('custom format');
-    expect(customFormatterInnerMock).toHaveBeenCalledTimes(1);
-    expect(customFormatterInnerMock).toHaveBeenCalledWith({
-      ...currencySettings,
-      value: customGBPMoney,
-      defaultFormatted: '£10.00',
+    describe('and the currency is registered', () => {
+      test('can be initialized with registered settings included', () => {
+        const currency = new Currency({ code: 'GBP', thousandsSeparator: '-' });
+        expect(currency.settings).toEqual({
+          ...GBP,
+          thousandsSeparator: '-',
+        });
+      });
     });
   });
 
-  test('"parser" option should do custom parsing', () => {
-    // wrapping the mock as the customParser gets frozen once passed in
-    const customParserInnerMock = jest.fn(() => 'custom parsed');
-    const customParser = (...args: any[]) => customParserInnerMock(...args);
-    const customGBP = new Currency({
-      ...Currency.getSettings('GBP'),
-      parser: customParser,
-    });
-    const parsedValue = customGBP.parse('£10.00');
-    const {
-      parser,
-      formatter,
-      ...currencySettings
-    } = customGBP.settings;
+  test('can be initialized with the static "init" factory function', () => {
 
-    expect(parsedValue).toBe('custom parsed');
-    expect(customParserInnerMock).toHaveBeenCalledTimes(1);
-
-    // "toHaveBeenCalledWith" cannot be used to the equality check on the "Money" instance
-    const call = customParserInnerMock.mock.calls[0];
-
-    expect(call.length).toBe(1);
-    expect(call[0]).toMatchObject({
-      ...currencySettings,
-      value: '£10.00',
-    });
-    expect(call[0].defaultParsed.equals(new Money('10.00', customGBP))).toBe(true);
+    const currency = Currency.init('USD');
+    expect(currency).toBeInstanceOf(Currency);
+    expect(currency.settings).toEqual(USD);
   });
 
-  test('new Currency("USD") === Currency.init("USD")', () => {
-    const
-      currencyA = new Currency('USD'),
-      currencyB = Currency.init('USD');
-    expect(currencyA.is(currencyB)).toBe(true);
-  });
+  describe('with valid values', () => {
+    const currency = new Currency('USD');
 
-  describe('when "code" is not provided', () => {
-    test('should throw an "InvalidCurrencyError"', () => {
-      expect(() => new Currency({
-        symbol: '£',
-      })).toThrow(new InvalidCurrencyError('Invalid currency settings; code is required.'));
+    test('should return a frozen object', () => {
+      expect(Object.isFrozen(currency)).toBe(true);
+    });
+
+    describe('"is"', () => {
+      describe('when given the same currency with the exact same settings', () => {
+        test('should return true', () => {
+          expect(currency.is('USD')).toBe(true);
+          expect(currency.is(new Currency('USD'))).toBe(true);
+        });
+      });
+
+      describe('when given a different currency', () => {
+        test('should return false', () => {
+          expect(currency.is('GBP')).toBe(false);
+          expect(currency.is(new Currency('GBP'))).toBe(false);
+        });
+      });
+
+      describe('when given the same currency with different settings', () => {
+        test('should return false', () => {
+          const customUSD = new Currency({
+            ...currency.settings,
+            symbol: 'US$',
+          });
+          expect(currency.is(customUSD)).toBe(false);
+        });
+      });
+    });
+
+    test('"toString()" should return the code', () => {
+      expect(currency.toString()).toBe('USD');
+    });
+
+    test('"toJSON()" should return the code', () => {
+      expect(currency.toJSON()).toBe('USD');
+    });
+
+    test(
+      '"clone()" should return a new instance of the Currency class with the same settings',
+      () => {
+        const clonedCurency = currency.clone();
+        expect(clonedCurency).toBeInstanceOf(Currency);
+        expect(clonedCurency).not.toBe(currency);
+        expect(currency.is(clonedCurency)).toBe(true);
+      },
+    );
+
+    test('"settings" getter should return the settings', () => {
+      expect(currency.settings).toEqual({
+        ...USD,
+        formatter: undefined,
+        parser: undefined,
+      });
     });
   });
 
-  describe('when invalid input is provided', () => {
-    test('should throw an "WrongInputError"', () => {
-      expect(() => new Currency(undefined as unknown as Currency))
-      .toThrow(new WrongInputError('Invalid currency provided.'));
+  describe('when given any invalid or missing code as part of input settings', () => {
+    const invalidCurrencyCodes: any[] = [
+      undefined,
+      null,
+      false,
+      true,
+      '',
+      123,
+      {},
+      [],
+    ];
+
+    test('"constructor" should throw an error', () => {
+      for (const invalidCurrencyCode of invalidCurrencyCodes) {
+        expect(() => new Currency({ code: invalidCurrencyCode as unknown as string }))
+          .toThrow(new InvalidCurrencyError('Invalid currency settings; code is required.'));
+      }
+    });
+  });
+
+  describe('when given any invalid input', () => {
+    const invalidCurrencyInputs: any[] = [
+      undefined,
+      null,
+      false,
+      true,
+      '',
+      123,
+      [],
+    ];
+
+    test('"constructor" should throw an error', () => {
+      for (const invalidCurrencyInput of invalidCurrencyInputs) {
+        expect(() => new Currency(invalidCurrencyInput as unknown as string))
+          .toThrow(new WrongInputError('Invalid currency provided.'));
+      }
     });
   });
 });
