+++ src/CurrencyStore/data.ts
@@ -7,4 +7,4 @@ export interface CurrencySettingsInternalStore {
 
 const data: CurrencySettingsInternalStore = {};
 
-export default data;
+export default () => data;
