+++ src/CurrencyStore/deregister.ts
@@ -1,10 +1,11 @@
-import data from './data';
+import getData from './data';
 
 /**
  * Delete a registered currency
  * @param code - Currency code
  */
 const deregister = (code: string) => {
+  const data = getData();
   return delete data[code.toUpperCase()];
 };
