+++ src/CurrencyStore/isRegistered.ts
@@ -1,10 +1,13 @@
-import data from './data';
+import getData from './data';
 
 /**
  * Check if currency has been registered
  * @param code - Currency code
  * @returns - True if currency has been registered
  */
-const isRegistered = (code: string) => code.toUpperCase() in data;
+const isRegistered = (code: string) => {
+  const data = getData();
+  return code.toUpperCase() in data;
+};
 
 export default isRegistered;
