+++ tsconfig.json
@@ -10,8 +10,7 @@
     "moduleResolution": "node",
     "allowSyntheticDefaultImports": true,
     "esModuleInterop": true,
-    "suppressImplicitAnyIndexErrors": true,
-    "strictPropertyInitialization": false
+    "suppressImplicitAnyIndexErrors": true
   },
   "include": [
     "src/**/*.ts"
